Write READMEs for `bin` and `test` directories.
I wrote simple READMEs for the `bin` and `test` directories, both because that hopefully explains what's happening there and because that allows me to successfully commit those directories so they'll be part of what people get when they fork/clone the project.

I also made a few minor changes to the top-level `README.md`.