Write tests for `create_username_dist.sh`.
This provides some simple `bats` acceptance tests for `create_username_dist.sh`. It's pretty all-or-nothing,
unfortunately, because it's hard to "reach in" and see how things are going part way through. (+1 squashed commit)
Squashed commits:
[783ef04] Fix `gmktemp` for Linux.

I had a leftover `gmktemp` from MacOS that needed to be changed to just `mktemp`
so it will work in the lab on Linux.

Write tests for `create_username_dist.sh`.

This provides some simple `bats` acceptance tests for `create_username_dist.sh`. It's pretty all-or-nothing,
unfortunately, because it's hard to "reach in" and see how things are going part way through.