Remove unnecessary order dependency in tests.
This removes all dependency on order in `process_logs.bats`
by sorting the generated HTML file and comparing that to a
sorted version of the target HTML file. This is a little
weird because you could, in theory, pass the test by
generating an HTML file with the lines in a random (and therefore
unusable) order. I figure that's not terribly likely, though,
and this makes it less likely that a legitimate change in the
ordering of the generated output will break a test.